refactor: move fleet-terraform-generator to tools
Move commands into tools/cmd/fleet-terraform-generator in preperation
for more commands.
