+++ .github/workflows/generate.yml
@@ -21,7 +21,7 @@ jobs:
       - uses: actions/setup-go@v5
         with:
           go-version: '1.22.x'
-          cache-dependency-path: fleet-terraform-generator/go.sum
+          cache-dependency-path: tools/go.sum
 
       - uses: actions/checkout@v4
         with:
