--- fleet-terraform-generator/docs/README.md
@@ -1,16 +1,0 @@
-## fleet-terraform-generator
-
-Tool for generating Terraform modules for managing Fleet integrations.
-
-### Options
-
-```
-  -h, --help   help for fleet-terraform-generator
-```
-
-### SEE ALSO
-
-* [fleet-terraform-generator generate](fleet-terraform-generator_generate.md)	 - Generate a Terraform module for managing a Fleet package policy.
-* [fleet-terraform-generator list](fleet-terraform-generator_list.md)	 - List all Fleet packages, policy templates, data streams, and inputs.
-
-
