--- fleet-terraform-generator/docs/fleet-terraform-generator_list.md
@@ -1,21 +1,0 @@
-## fleet-terraform-generator list
-
-List all Fleet packages, policy templates, data streams, and inputs.
-
-```
-fleet-terraform-generator list [flags]
-```
-
-### Options
-
-```
-      --continue-on-error     Continue on errors related to reading package specifications.
-  -h, --help                  help for list
-      --packages-dir string   Directory containing Fleet packages.
-```
-
-### SEE ALSO
-
-* [fleet-terraform-generator](fleet-terraform-generator.md)	 - Tool for generating Terraform modules for managing Fleet integrations.
-
-
