+++ tools/cmd/fleet-terraform-generator/main.go
@@ -15,8 +15,16 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package docs
+package main
 
-//go:generate go run ../internal/docs
-//go:generate sh -c "perl -i -pe 's/^###### Auto generated by spf13.*$//g;' *.md"
-//go:generate mv fleet-terraform-generator.md README.md
+import (
+	"os"
+
+	"github.com/elastic/terraform-module-fleet/tools/internal/cmd"
+)
+
+func main() {
+	if err := cmd.RootCmd().Execute(); err != nil {
+		os.Exit(1)
+	}
+}
